Change parent node of supernode from 0 to null in database and file format fixes
